bugfix: always write same size microscope_image
When microscope_image is written to main.microscope_image,
use a 480x640 array of zeroes when the microscope camera
is not active. This fixes a size mismatch error that occured
when switching the microscope camera on and off while
main.microscope_image is being collected.

Add a try...catch block to part of the fasttimerFcn to help
with future debugging.